backend/requirements.txt
fastapi==0.126.0
uvicorn==0.38.0

# Market data + transport
alpaca-py==0.43.2
websockets
httpx
requests
tenacity
google-genai

# Firebase / Firestore
firebase-admin
google-cloud-firestore
google-genai

# Runtime config / messaging / utilities
python-dotenv
nats-py
pandas

# Optional DB drivers (Postgres)
psycopg[binary]
asyncpg==0.29.0
psycopg2-binary==2.9.9
